コンテキストメニューの削除ボタンをコンポーネント化
