+++ app/components/folder-explorer/FolderLink.tsx
@@ -1,64 +1,11 @@
-import {
-  IconFolderFilled,
-  IconTrash,
-  IconEdit,
-  IconHelp,
-  IconExternalLink,
-  IconLogin2
-} from "@tabler/icons-react"
-import { ActionIcon, HoverCard, Menu, UnstyledButton, Text, Group, Stack } from "@mantine/core"
+import { IconFolderFilled, IconEdit, IconLogin2 } from "@tabler/icons-react"
+import { Menu, UnstyledButton } from "@mantine/core"
 import { clsx } from "clsx"
 import styles from "./EntryLink.module.css"
-import { useState } from "react"
-import { Link } from "react-router"
-import { useEmptyFolderDeleteUi } from "~/usecases/folder-explorer/delete/ui.hooks"
-import { notifications } from "@mantine/notifications"
-import { errorContent, successContent } from "~/libs/mantine-notifications/options"
-import IconLoadingSpinner from "../icon-loading-spinner/IconLoadingSpinner"
 import EntryCheckbox from "./EntryCheckbox"
 import type { Entry } from "~/usecases/folder-explorer/types"
-
-function DeleteDisabledHelp() {
-  return (
-    <HoverCard
-      shadow="md"
-      withArrow
-      arrowPosition="center"
-      position="right"
-      offset={0}
-      styles={{
-        arrow: { "--popover-border-color": "var(--mantine-color-pale-indigo-2)" },
-        dropdown: { "--popover-border-color": "var(--mantine-color-pale-indigo-2)" }
-      }}
-    >
-      <HoverCard.Target>
-        <ActionIcon variant="white" color="gray" radius="xl" aria-label="help">
-          <IconHelp size={16} />
-        </ActionIcon>
-      </HoverCard.Target>
-      <HoverCard.Dropdown>
-        <Stack gap={6}>
-          <Text size="xs">空でないフォルダは削除できません。</Text>
-          <Link
-            to="/folder-map"
-            reloadDocument
-            target="_blank"
-            rel="noopener noreferrer"
-            style={{
-              textDecoration: "none",
-              color: "var(--mantine-color-pink-6)"
-            }}
-          >
-            <Group gap={4} justify="end" align="center">
-              <Text size="xs">フォルダ構成を編集</Text>
-              <IconExternalLink size={14} color="var(--mantine-color-pink-5)" />
-            </Group>
-          </Link>
-        </Stack>
-      </HoverCard.Dropdown>
-    </HoverCard>
-  )
-}
+import MenuItemForDelete from "./folder-context-menu/MenuItemForDelete"
+import { useState } from "react"
 
 interface Props {
   folder: { id: number; name: string }
@@ -77,10 +24,8 @@ export default function FolderLink({
   selectable,
   updateSelection
 }: Props) {
-  const [openedMenu, setOpenedMenu] = useState(false)
+  const [isOpenedContextMenu, setIsOpenedContextMenu] = useState(false)
   const isEmpty = folderEntryCount === 0
-
-  const { deleteFolder, isDeleting } = useEmptyFolderDeleteUi()
 
   return selectable ? (
     <div
@@ -105,8 +50,8 @@ export default function FolderLink({
       position="right-start"
       withArrow
       arrowPosition="center"
-      opened={openedMenu}
-      onChange={setOpenedMenu}
+      opened={isOpenedContextMenu}
+      onChange={setIsOpenedContextMenu}
       styles={{
         arrow: { "--popover-border-color": "var(--mantine-color-pale-indigo-2)" },
         dropdown: { "--popover-border-color": "var(--mantine-color-pale-indigo-2)" },
@@ -118,13 +63,13 @@ export default function FolderLink({
           onClick={onLinkClick}
           onContextMenu={(e) => {
             e.preventDefault()
-            setOpenedMenu(true)
+            setIsOpenedContextMenu(true)
           }}
           className={clsx(
             styles.entry_link,
             styles.folder_link,
             isActiveStyle && styles.highlight_active,
-            openedMenu && styles.highlight_focused
+            isOpenedContextMenu && styles.highlight_focused
           )}
         >
           <IconFolderFilled size={18} />
@@ -137,27 +82,11 @@ export default function FolderLink({
         <Menu.Item leftSection={<IconEdit size={16} />}>フォルダ名を変更</Menu.Item>
         <Menu.Item leftSection={<IconLogin2 size={16} />}>ここに移動するアイテムを選択</Menu.Item>
         <Menu.Divider />
-        <Menu.Item
-          color="red"
-          leftSection={isDeleting ? <IconLoadingSpinner size={16} /> : <IconTrash size={16} />}
-          rightSection={!isEmpty && <DeleteDisabledHelp />}
-          disabled={!isEmpty || isDeleting}
-          component={isEmpty ? "button" : "div"}
-          onClick={() => {
-            if (!isEmpty) return
-            deleteFolder(folder, {
-              onSuccess: () => {
-                setOpenedMenu(false)
-                notifications.show(successContent(`フォルダ「${folder.name}」を削除しました`))
-              },
-              onError: ({ detail }) => {
-                notifications.show(errorContent(detail.message, detail.target))
-              }
-            })
-          }}
-        >
-          フォルダを削除
-        </Menu.Item>
+        <MenuItemForDelete
+          isEmpty={isEmpty}
+          folder={folder}
+          setIsOpenedContextMenu={setIsOpenedContextMenu}
+        />
       </Menu.Dropdown>
     </Menu>
   )
